requirements.txt
python-dotenv~=1.0.1
setuptools~=58.1.0
discord~=2.3.2